Added MRU_Open_File_Reuse_Current_Empty_Tab option
By default, while opening a file in a tab, if the current tab is
completely empty, it tries to find a tab which contains the required
file. If it does not find it, it creates a new tab for it.
However, if the current tab is completely empty, we can re-use it.