Keep alternate file when using current window
Update commands to use keepalt when MRU set to use the current window,
such that after opening and selecting a file from the MRU window, the
alternate file is the last edited file rather than the MRU files buffer.

The MRU files buffer is empty when loaded as the alternate file, so it
doesn't seem necessary to add a new configuration variable for this.