Fix MRUToggle command when using current window
When loading MRU in the current window, toggling should not close the
window, but go back to the previous buffer. As there are circumstances
in which the MRU buffer will open in a split even when set to use the
current window, check for the previous buffer before switching back.