Tweak toggle and alt file behaviour when MRU set to use current window (#51)
* Wipe MRU buffer on hide when using current window

This avoids ever setting the MRU files buffer as the alternate file,
which prevents switching back to an empty MRU files list when MRU is
the first buffer loaded after opening vim. Prior to this, opening vim,
running the MRU command, selecting a file, and then switching to the
alternate file using CTRL-^ would load up an empty MRU files buffer.

* Fix MRUToggle command when using current window

When loading MRU in the current window, toggling should not close the
window, but go back to the previous buffer. As there are circumstances
in which the MRU buffer will open in a split even when set to use the
current window, check for the previous buffer before switching back.