Do not useopen when MRU set to use current window
When MRU is set to open in the current window, open a selected file
in the current window, even if it is already open in another window.

Note that this matches the behaviour of other plugins like Netrw and
bufexplorer when they are configured to open in the current window.

This commit does not change the behaviour for opening multiple files,
though it also doesn't add any tests to assert that, maybe it should?