Use global fzf window layout for FZFMru command
It seems odd to load FZFMru with a different layout to other FZF
commands, so remove the down option override when calling fzf#wrap.

I think I understand why this was added, it makes the FZFMru window the
same layout as other MRU windows (when not set to use current window),
but this really isn't an MRU window, it's a FZF window with MRU files.