Add option to open file using relative path
MRU normally opens files using the path found in the MRU list, which by
default shows the full path to the file. For people using the default
Vim statusline or using %f in a custom statusline, this means a file
opened via MRU will never been shown relative to the current directory.

This commit adds an option to modify the path before opening the file so
it will be opened relative to the home directory or current directory if
possible, which is also consistent with how the FZFMru command works.

There are other options here, it is possible to use MRU_Filename_Format
to show relative paths, but I think having an option to open files using
relative paths is much neater (the default display in the MRU files list
is very nice as it is, showing relative paths here can get a bit fugly).